fix: Avoid changing the URL of the service worker.
According to the official documentation[1], it's bad practice to change
the URL of the service worker. Also there's not really a point in having
time stamped URLs for the sw-register.js script. This also allows to
simplify the registration quite a bit, leading to less JavaScript to
parse and execute for the browser.

[1] https://developers.google.com/web/fundamentals/primers/service-workers/lifecycle